requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
pyarrow>=14.0.0
rich>=13.0.0
python-dateutil>=2.8.2
